refactor(cache): drop remote cache, keep it local-only
Remove turboCache (Turborepo Remote Cache), the tieredCache combinator, and
the Artifact serializer that only existed for remote transport. The package
now ships fsCache (local disk) and memoryCache, which covers the core
incremental-build use case with a much smaller surface.

Rename the exported helper classes to PascalCase files to match the
Telemetry convention: Fingerprint (core) and Manifest (@kubb/cache).
