+++ packages/cache/src/Manifest.ts
