+++ packages/core/src/Fingerprint.test.ts
@@ -4,7 +4,7 @@ import { join } from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
 import type { AdapterSource } from './createAdapter.ts'
 import type { Config } from './createKubb.ts'
-import { Fingerprint } from './fingerprint.ts'
+import { Fingerprint } from './Fingerprint.ts'
 
 function makeConfig(overrides: Partial<Config> = {}): Config {
   return {
